+++ src/features/residents/index.ts
@@ -1,6 +1,9 @@
 export * from './types';
 export * from './data/mockResidents';
 export * from './utils/formatters';
+export * from './services/residentService';
+export * from './hooks/useResidents';
+export * from './hooks/useResident';
 export { default as ResidentsPage } from './pages/ResidentsPage';
 export { default as ResidentProfilePage } from './pages/ResidentProfilePage';
 export { default as ResidentOnboardingPage } from './pages/ResidentOnboardingPage';